use-cases/curious-agent/requirements.txt
pydantic
google-generativeai
python-dotenv
